Move BaseCRUD tests to test_base_crud
Signed off by: Diego Araújo <diegoamc90@gmail.com>
